pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "content_extraction"
version = "0.1.0"
description = "Project dedicated to content extraction from unstructured files that contain some useful information."
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["beautifulsoup4>=4.13.4", "lxml>=6.0.0", "python-pptx>=1.0.2"]

[dependency-groups]
dev = [
    "dspy>=2.6.27",
    "jupyterlab>=4.4.5",
    "pre-commit>=4.2.0",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
builtins = ["_"]
target-version = "py312"
extend-exclude = [
    '.git',
    '__pycache__',
    'build',
    'dist',
    '.venv',
    'venv',
    '.tox',
    '.mypy_cache',
    '.pytest_cache',
    '*/migrations',
    '*/dev-tools',
]
line-length = 120

[tool.ruff.format]
quote-style = "single"
